Centralize build configuration
