+++ src/Components/CustomElements/src/js/tsconfig.json
@@ -1,9 +1,3 @@
 {
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "lib": [ "DOM", "ES2022" ],
-    "strict": true,
-    "forceConsistentCasingInFileNames": true
-  }
+  "extends": "../../../Shared.JS/tsconfig.json",
 }
